+++ functions.php
@@ -9,4 +9,4 @@ require_once get_template_directory() . '/inc/posts.php';
 require_once get_template_directory() . '/inc/volumes.php';
 require_once get_template_directory() . '/inc/tags.php';
 require_once get_template_directory() . '/inc/more-link.php';
-
+require_once get_template_directory() . '/inc/cz-continue-reading.php';