+++ page-logout.php
@@ -4,4 +4,3 @@ if (is_user_logged_in()) {
 	wp_redirect(home_url());
 	exit;
 }
-?>