+++ pages/page-home.php
@@ -7,6 +7,7 @@
 </section>
 
 <?php get_template_part('patterns/template-home-articles'); ?>
+<?php get_template_part('patterns/template-home-readings'); ?>
 <?php get_template_part('patterns/template-home-authors'); ?>
 <?php get_template_part('patterns/template-home-volumes'); ?>
 <?php get_template_part('patterns/template-home-categories'); ?>