+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `404.php` page
 - Add `inc/authentication.php` to manage user registration email verification, hide the WordPress top bar and disallow accessing the /dashboard to non admin users.
+### Changed
+- Update Home Page to supper CZ Continue Reading Plugin
+- Now users can register and are immediately logged in
 ### Fixed
 - Fix successfull login redirect
 - Fix `.success-message` appearance